Add script to initialize a Nitrokey HOTP secret to specific counter
Normally when one sets up the Nitrokey HOTP secret the assumption is
that the counter starts at zero. However in the case of using a counter
from a TPM, new counters are initialized at the value of the last
counter plus one. This means we need to synchronize the Nitrokey counter
with the TPM counter before we can use it.

This script takes advantage of the fact that the Nitrokey accepts the
next ten HOTP values in a list and then sets the Nitrokey counter to
that value, so the script "hops" ahead up until the point that the
Nitrokey counter matches the specified counter.