Check secrets app PIN before setting the new reverse hotp credential
If the PIN is not set, configure it.

Nitrokeys prior to https://github.com/Nitrokey/trussed-secrets-app/pull/114
will ignore the PIN. It might surprise users that have not set a PIN
in the past that a PIN is now set.

Nitrokeys with firmware after https://github.com/Nitrokey/trussed-secrets-app/pull/114
will accept expect this behaviour from hotp-verification